requirements.txt
fontawesome==5.10.1.post1
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
plotly==5.9.0
streamlit==1.21.0
streamlit_authenticator==0.2.2
